[add] #13 프로젝트 구조 개편, connection test용 controller 생성
