+++ src/main/java/com/example/exitmedserver/TestController.java
@@ -1,0 +1,14 @@
+package com.example.exitmedserver;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+public class TestController {
+    @GetMapping("/test")
+    @ResponseBody
+    public String test() {
+        return "test success!";
+    }
+}
