+++ src/main/java/com/example/exitmedserver/pill/repository/PillRepository.java
@@ -1,5 +1,6 @@
-package com.example.exitmedserver.pill;
+package com.example.exitmedserver.pill.repository;
 
+import com.example.exitmedserver.pill.entity.Pill;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
