+++ src/main/java/com/example/exitmedserver/pill/service/PillService.java
@@ -1,4 +1,4 @@
-package com.example.exitmedserver.pill;
+package com.example.exitmedserver.pill.service;
 
 import org.springframework.stereotype.Service;
