+++ src/main/java/com/example/exitmedserver/search/controller/SearchController.java
@@ -1,4 +1,4 @@
-package com.example.exitmedserver.search;
+package com.example.exitmedserver.search.controller;
 
 import org.springframework.stereotype.Controller;
