+++ src/main/java/com/example/exitmedserver/search/entity/FavoriteList.java
@@ -1,22 +1,17 @@
-package com.example.exitmedserver.search;
+package com.example.exitmedserver.search.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.Getter;
 
-import java.sql.Timestamp;
-
 @Entity
 @Getter
-public class SearchHistoryList {
+public class FavoriteList {
     @Id
     @Column(name = "user_id")
     private String userId;
 
     @Column(name = "pill_item_sequence")
     private Long pillItemSequence;
-
-    @Column(name = "created_at")
-    private Timestamp createdAt;
 }
