+++ src/main/java/com/example/exitmedserver/search/repository/FavoriteListRepository.java
@@ -1,5 +1,6 @@
-package com.example.exitmedserver.search;
+package com.example.exitmedserver.search.repository;
 
+import com.example.exitmedserver.search.entity.FavoriteList;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
