--- src/main/java/com/example/exitmedserver/user/UserController.java
@@ -1,7 +1,0 @@
-package com.example.exitmedserver.user;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class UserController {
-}
