+++ src/main/java/com/example/exitmedserver/user/controller/UserController.java
@@ -1,0 +1,8 @@
+package com.example.exitmedserver.user.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+
+}
