+++ src/main/java/com/example/exitmedserver/user/dto/UserSignupRequestDto.java
@@ -1,0 +1,24 @@
+package com.example.exitmedserver.user.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+
+@Data
+public class UserSignupRequestDto {
+
+    @NotNull
+    private String userId;
+
+    @NotNull
+    private String userPassword;
+
+    @NotNull
+    private String fullName;
+
+    @NotNull
+    private int dateOfBirth;
+
+    private String sex;
+
+    private boolean isPregnant;
+}
