+++ src/main/java/com/example/exitmedserver/user/repository/UserProfileRepository.java
@@ -1,5 +1,6 @@
-package com.example.exitmedserver.user;
+package com.example.exitmedserver.user.repository;
 
+import com.example.exitmedserver.user.entity.UserProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
