+++ src/main/java/com/example/exitmedserver/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.example.exitmedserver.user;
+package com.example.exitmedserver.user.repository;
 
+import com.example.exitmedserver.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
